Added the agent and part of the algorithm test
the agent appears to be working fine now, the movement actions are working.

testing the agents movement in the maze without any training looks to be working as well. However, I still need to optimise it to ensure the sub goal part works and place all the comments in a more tasteful manner.