Animated agent maze traversal
Not too important for the final hand-in. But still cool to have. When running AnimatedTestAgent.py you can see the agent actually going through the maze and the steps it takes. However, it is really slow and messy because it is untrained. Repeating this process for the trained version would be cool, too.